requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.6
django-ckeditor==6.0.0
django-js-asset==1.2.2
filelock==3.0.12
Pillow==8.0.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
virtualenv==20.2.2
virtualenvwrapper-win==1.2.6
